Common/util.py
import logging
import os
import sys
from datetime import datetime, timedelta
from time import sleep
import azure.functions as func
from azure.storage.blob import BlobServiceClient, BlobClient, ContainerClient,generate_blob_sas, BlobSasPermissions

#Copy blob from source container to destination container
def copy_blob():
    pass 
